requirements.txt
asgiref==3.3.1
Django==3.1.6
django-filter==2.4.0
djangorestframework==3.12.2
Markdown==3.3.3
pytz==2021.1
sqlparse==0.4.1
psycopg2-binary>=2.8
